ModulePreservation

Computes module preservation statistics in a query dataset for a given reference dataset

Usage

ModulePreservation(
  seurat_obj,
  seurat_ref,
  name,
  n_permutations = 500,
  parallel = FALSE,
  seed = 12345,
  return_raw = FALSE,
  wgcna_name = NULL,
  wgcna_name_ref = NULL,
  ...
)

Arguments

seurat_obj

A Seurat object

seurat_ref

A Seurat object serving as the reference for the module preservation analysis

name

The name to give the module preservation analysis.

n_permutations

Number of permutations for the module preservation test.

parallel

logical determining whether to run preservation analysis in parallel

seed

random seed for the permutation analysis.

return_raw

if TRUE, returns the module preservation statistics, else returns seurat_obj with the stats added to the scWGCNA experiment.

wgcna_name

The name of the scWGCNA experiment in the seurat_obj@misc slot

wgcna_name_ref

The name of the scWGCNA experiment in the seurat_ref@misc slot

Examples

ModulePreservation
#> function (seurat_obj, seurat_ref, name, n_permutations = 500, 
#>     parallel = FALSE, seed = 12345, return_raw = FALSE, wgcna_name = NULL, 
#>     wgcna_name_ref = NULL, ...) 
#> {
#>     if (is.null(wgcna_name)) {
#>         wgcna_name <- seurat_obj@misc$active_wgcna
#>     }
#>     if (is.null(wgcna_name_ref)) {
#>         wgcna_name_ref <- seurat_ref@misc$active_wgcna
#>     }
#>     print("Setup datasets")
#>     datExpr_ref <- GetDatExpr(seurat_ref, wgcna_name_ref)
#>     datExpr_query <- GetDatExpr(seurat_obj, wgcna_name)
#>     setLabels <- c("ref", "query")
#>     multiExpr <- list(ref = list(data = datExpr_ref), query = list(data = datExpr_query))
#>     ref_modules <- list(ref = GetModules(seurat_ref)$module)
#>     print("ref:")
#>     print(dim(datExpr_ref))
#>     print("query:")
#>     print(dim(datExpr_query))
#>     print("Run Module Preservation")
#>     mp <- WGCNA::modulePreservation(multiExpr, ref_modules, referenceNetworks = 1, 
#>         nPermutations = n_permutations, randomSeed = seed, quickCor = 0, 
#>         parallelCalculation = parallel, ...)
#>     if (return_raw) {
#>         return(mp)
#>     }
#>     ref <- 1
#>     test <- 2
#>     statsObs <- cbind(mp$quality$observed[[ref]][[test]], mp$preservation$observed[[ref]][[test]])
#>     statsZ <- cbind(mp$quality$Z[[ref]][[test]], mp$preservation$Z[[ref]][[test]])
#>     mod_pres <- list(obs = statsObs, Z = statsZ)
#>     seurat_obj <- SetModulePreservation(seurat_obj, mod_pres, 
#>         name, wgcna_name)
#>     seurat_obj
#> }
#> <bytecode: 0x7faa5882d900>
#> <environment: namespace:scWGCNA>