ModuleTraitCorrelation'

Correlates categorical and numeric variables with Module Eigengenes or hub-gene scores.

Usage

ModuleTraitCorrelation(
  seurat_obj,
  traits,
  group.by = NULL,
  features = "hMEs",
  cor_method = "pearson",
  wgcna_name = NULL,
  ...
)

Arguments

seurat_obj

A list of column names in the Seurat object's metadata that you wish to correlate with each module. Traits must be a categorical variable (not a character vector), or a numeric variable.

features

Which features to use to summarize each modules? Valid choices are hMEs, MEs, or scores

cor_meth

Which method to use for correlation? Valid choices are pearson, spearman, kendall.

Examples

ModuleTraitCorrelation
#> function (seurat_obj, traits, group.by = NULL, features = "hMEs", 
#>     cor_method = "pearson", wgcna_name = NULL, ...) 
#> {
#>     if (is.null(wgcna_name)) {
#>         wgcna_name <- seurat_obj@misc$active_wgcna
#>     }
#>     if (features == "hMEs") {
#>         MEs <- GetMEs(seurat_obj, TRUE, wgcna_name)
#>     }
#>     else if (features == "MEs") {
#>         MEs <- GetMEs(seurat_obj, FALSE, wgcna_name)
#>     }
#>     else if (features == "scores") {
#>         MEs <- GetModuleScores(seurat_obj, wgcna_name)
#>     }
#>     else {
#>         stop("Invalid feature selection. Valid choices: hMEs, MEs, scores, average")
#>     }
#>     if (sum(traits %in% colnames(seurat_obj@meta.data)) != length(traits)) {
#>         stop(paste("Some of the provided traits were not found in the Seurat obj:", 
#>             paste(traits[!(traits %in% colnames(seurat_obj@meta.data))], 
#>                 collapse = ", ")))
#>     }
#>     if (is.null(group.by)) {
#>         group.by <- "temp_ident"
#>         seurat_obj$temp_ident <- Idents(seurat_obj)
#>     }
#>     valid_types <- c("numeric", "factor", "integer")
#>     data_types <- sapply(traits, function(x) {
#>         class(seurat_obj@meta.data[, x])
#>     })
#>     if (!all(data_types %in% valid_types)) {
#>         incorrect <- traits[!(data_types %in% valid_types)]
#>         stop(paste0("Invalid data types for ", paste(incorrect, 
#>             collapse = ", "), ". Accepted data types are numeric, factor, integer."))
#>     }
#>     if (any(data_types == "factor")) {
#>         factor_traits <- traits[data_types == "factor"]
#>         for (tr in factor_traits) {
#>             warning(paste0("Trait ", tr, " is a factor with levels ", 
#>                 paste0(levels(seurat_obj@meta.data[, tr]), collapse = ", "), 
#>                 ". Levels will be converted to numeric IN THIS ORDER for the correlation, is this the expected order?"))
#>         }
#>     }
#>     modules <- GetModules(seurat_obj, wgcna_name)
#>     mods <- levels(modules$module)
#>     mods <- mods[mods != "grey"]
#>     trait_df <- seurat_obj@meta.data[, traits]
#>     if (any(data_types == "factor")) {
#>         factor_traits <- traits[data_types == "factor"]
#>         for (tr in factor_traits) {
#>             trait_df[, tr] <- as.numeric(trait_df[, tr])
#>         }
#>     }
#>     cor_list <- list()
#>     cor_list[["all_cells"]] <- cor(as.matrix(trait_df), as.matrix(MEs), 
#>         method = cor_method)
#>     trait_df <- cbind(trait_df, seurat_obj@meta.data[, group.by])
#>     colnames(trait_df)[ncol(trait_df)] <- "group"
#>     MEs <- cbind(as.data.frame(MEs), seurat_obj@meta.data[, group.by])
#>     colnames(MEs)[ncol(MEs)] <- "group"
#>     if (class(seurat_obj@meta.data[, group.by]) == "factor") {
#>         group_names <- levels(seurat_obj@meta.data[, group.by])
#>     }
#>     else {
#>         group_names <- levels(as.factor(seurat_obj@meta.data[, 
#>             group.by]))
#>     }
#>     trait_list <- dplyr::group_split(trait_df, group, .keep = FALSE)
#>     ME_list <- dplyr::group_split(MEs, group, .keep = FALSE)
#>     names(trait_list) <- group_names
#>     names(ME_list) <- group_names
#>     for (i in names(trait_list)) {
#>         cor_list[[i]] <- cor(as.matrix(trait_list[[i]]), as.matrix(ME_list[[i]]), 
#>             method = cor_method)
#>     }
#>     mt_cor <- list(cor_mat = cor_list, pval = NA, fdr = NA)
#>     seurat_obj <- SetModuleTraitCorrelation(seurat_obj, mt_cor, 
#>         wgcna_name)
#>     seurat_obj
#> }
#> <bytecode: 0x7fe6cd00f7b0>
#> <environment: namespace:scWGCNA>