TransferModuleGenome

Takes a module table and a gene mapping table (like from biomart) with gene names from two genomes in order to switch

Usage

TransferModuleGenome(
  modules,
  gene_mapping,
  genome1_col = NULL,
  genome2_col = NULL
)

Arguments

modules

module table from GetModules function

gene_mapping

table that has the gene names for both genomes

genome1_col

the column in gene_mapping that has gene names for the genome currently present in modules

genome2_col

the column in gene_mapping that has gene names for the genome to transfer to

Examples

TransferModuleGenome
#> function (modules, gene_mapping, genome1_col = NULL, genome2_col = NULL) 
#> {
#>     if (is.null(genome1_col)) {
#>         genome1_col <- colnames(gene_mapping)[1]
#>     }
#>     if (is.null(genome2_col)) {
#>         genome2_col <- colnames(gene_mapping)[2]
#>     }
#>     gene_mapping <- gene_mapping[, c(genome1_col, genome2_col)]
#>     gene_list <- modules$gene_name
#>     gene_mapping <- gene_mapping[gene_mapping[, genome1_col] %in% 
#>         gene_list, ]
#>     modules <- subset(modules, gene_name %in% gene_mapping[, 
#>         genome1_col])
#>     print("here")
#>     gene_match <- match(gene_list, gene_mapping[, genome1_col])
#>     gene_mapping <- na.omit(gene_mapping[gene_match, ])
#>     print("here")
#>     print(dim(modules))
#>     print(dim(gene_mapping))
#>     print(length(gene_match))
#>     modules$gene_name <- gene_mapping[, genome2_col]
#>     modules
#> }
#> <bytecode: 0x7f9018a3bdb8>
#> <environment: namespace:scWGCNA>